renamed user_discovery to update_shotgun_users
